basic layout concept for 8 SPI interfaces on one FMC carrier
using existing parts, but utilizing both sides of the PCB
only added 4 ports for now to evaluate the basic stragety
there are a lot of traces that need to be split etc.
will have to switch to flatter DC-DC for this
also check if we can use 2 power supplies for 8 posrts or if we need to bump up the current capacity.
might be able to expand the back of the pcb outline behind the FMC a bit and stay within the vita spec?